requirements.txt
argon2-cffi==16.2.0
cffi==1.8.3
Django==1.10.2
django-cors-headers==1.2.2
django-filter==0.15.2
djangorestframework==3.4.7
djoser==0.5.1
gunicorn==19.6.0
httpie==0.9.6
Markdown==2.6.7
pycparser==2.16
Pygments==2.1.3
requests==2.11.1
six==1.10.0
